consolidate all global initialization into globals.go to avoid init order dependency issues
